// frontend/src/app/Modulos/ventas/pages/cilindros-cliente/components/Interface/CilindroCliente.interface.ts
export interface CilindroCliente {
    CVECLIENTE_C:   number;
    NOMBRE_C:       string;
    FHEMISION:      Date;
    CVEPRODUCTO_CP: string;
    CVEENVASE_F:    string;
    DESCENVASE:     string;
    DESCCORTA_CP:   string;
    IDCILINDRO:     number;
    SERIE:          string;
    NUMREMISION:    number;
    SUCURSAL:       string;
}